Omit unneeded checks in validation.
There is nothing that says container_no should be a String
in the specs. If it's not a String (but maybe an unknown
Container object), we can request the `to_s` value and still
validate a container number.

Also, the regular expression, together with the `to_s` will
handle the nil and empty string cases.